Update beamlines to use shared
