+++ src/dodal/beamlines/b07_shared.py
@@ -2,16 +2,17 @@ from dodal.common.beamlines.beamline_utils import (
     device_factory,
 )
 from dodal.common.beamlines.beamline_utils import set_beamline as set_utils_beamline
-from dodal.devices.synchrotron import Synchrotron
+from dodal.devices.b07 import Grating
+from dodal.devices.pgm import PGM
 from dodal.log import set_beamline as set_log_beamline
 from dodal.utils import BeamlinePrefix, get_beamline_name
 
-BL = get_beamline_name("i09-2")
-PREFIX = BeamlinePrefix(BL, suffix="J")
+BL = get_beamline_name("b07")
+PREFIX = BeamlinePrefix(BL, suffix="B")
 set_log_beamline(BL)
 set_utils_beamline(BL)
 
 
 @device_factory()
-def synchrotron() -> Synchrotron:
-    return Synchrotron()
+def pgm() -> PGM:
+    return PGM(prefix=f"{PREFIX.beamline_prefix}-OP-PGM-01:", grating=Grating)
