+++ src/dodal/beamlines/i05_1.py
@@ -1,7 +1,5 @@
-from dodal.beamline_specific_utils.i05_shared import pgm as i05_pgm
 from dodal.common.beamlines.beamline_utils import device_factory
 from dodal.common.beamlines.beamline_utils import set_beamline as set_utils_beamline
-from dodal.devices.pgm import PGM
 from dodal.devices.synchrotron import Synchrotron
 from dodal.log import set_beamline as set_log_beamline
 from dodal.utils import BeamlinePrefix, get_beamline_name
@@ -10,11 +8,6 @@ BL = get_beamline_name("i05-1")
 PREFIX = BeamlinePrefix(BL, suffix="J")
 set_log_beamline(BL)
 set_utils_beamline(BL)
-
-
-@device_factory()
-def pgm() -> PGM:
-    return i05_pgm()
 
 
 @device_factory()
