+++ src/dodal/beamlines/i09.py
@@ -1,11 +1,14 @@
+from dodal.beamlines.i09_1_shared import dcm
+from dodal.beamlines.i09_2_shared import pgm
 from dodal.common.beamlines.beamline_utils import (
     device_factory,
 )
 from dodal.common.beamlines.beamline_utils import set_beamline as set_utils_beamline
 from dodal.devices.electron_analyser import SelectedSource
 from dodal.devices.electron_analyser.vgscienta import VGScientaAnalyserDriverIO
-from dodal.devices.i09 import DCM, Grating, LensMode, PassEnergy, PsuMode
-from dodal.devices.pgm import PGM
+from dodal.devices.i09 import LensMode, PassEnergy, PsuMode
+
+# from dodal.devices.pgm import PGM
 from dodal.devices.synchrotron import Synchrotron
 from dodal.log import set_beamline as set_log_beamline
 from dodal.utils import BeamlinePrefix, get_beamline_name
@@ -21,21 +24,8 @@ def synchrotron() -> Synchrotron:
     return Synchrotron()
 
 
-@device_factory()
-def pgm() -> PGM:
-    return PGM(
-        prefix=f"{BeamlinePrefix(BL, suffix='J').beamline_prefix}-MO-PGM-01:",
-        grating=Grating,
-    )
-
-
-@device_factory()
-def dcm() -> DCM:
-    return DCM(prefix=f"{PREFIX.beamline_prefix}-MO-DCM-01:")
-
-
-# Connect will work again after this work completed
-# https://jira.diamond.ac.uk/browse/I09-651
+# # Connect will work again after this work completed
+# # https://jira.diamond.ac.uk/browse/I09-651
 @device_factory()
 def analyser_driver() -> VGScientaAnalyserDriverIO[LensMode, PsuMode, PassEnergy]:
     energy_sources = {
