+++ src/dodal/beamlines/i09_1.py
@@ -1,10 +1,10 @@
+from dodal.beamlines.i09_1_shared import dcm
 from dodal.common.beamlines.beamline_utils import (
     device_factory,
 )
 from dodal.common.beamlines.beamline_utils import set_beamline as set_utils_beamline
 from dodal.devices.electron_analyser import SelectedSource
 from dodal.devices.electron_analyser.specs import SpecsAnalyserDriverIO
-from dodal.devices.i09.dcm import DCM
 from dodal.devices.i09_1 import LensMode, PsuMode
 from dodal.devices.synchrotron import Synchrotron
 from dodal.log import set_beamline as set_log_beamline
@@ -19,11 +19,6 @@ set_utils_beamline(BL)
 @device_factory()
 def synchrotron() -> Synchrotron:
     return Synchrotron()
-
-
-@device_factory()
-def dcm() -> DCM:
-    return DCM(prefix=f"{PREFIX.beamline_prefix}-MO-DCM-01:")
 
 
 # Connect will work again after this work completed
